add app tests and uc validations

+++ src/applications/DeleteCar.ts
@@ -4,6 +4,9 @@ export default class DeleteCar {
   constructor(private repo: CarRepo){}
 
   execute(id: string): boolean{
+    if(!id)
+      throw new Error("Parameter id is missing!")
+
     const car = this.repo.delete(id)
 
     if(!car)
